python_predictor/requirements.txt
ultralytics
torch
torchvision
Pillow
Flask
Flask-Cors
opencv-python-headless